scx_rusty: Use bpf_skel::types::* instead of bpf_intf::*
libbpf-rs now provides all types defined in .bpf.c. Use those definitions
instead of bindgen generated ones. This removes the need for explicit type
casts.
